requirements.txt
flask==2.0.3
Werkzeug==2.0.3
numpy==1.21.6
opencv-python==4.5.5.64
pillow==9.0.1
flask-cors==3.0.10 